fix: remove action and channel from Operation Trait Object
